refactor: split large agent/team/api modules into subfiles
